requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
PyYAML>=6.0.1
Jinja2>=3.1.4
requests>=2.32.3
litellm>=1.35.7,<1.77
jsonschema>=4.22.0
jmespath>=1.0.1